+++ compiler/build.gradle
@@ -37,9 +37,9 @@ dependencies {
     implementation(project(":ownerContext"))
     implementation(project(":translator"))
     implementation(project(":translator-string"))
-    implementation(project(":idl:idl-declarations"))
-    implementation(project(":idl:idl-parser"))
-    implementation(project(":idl:idl-models"))
+    implementation(project(":idl-declarations"))
+    implementation(project(":idl-parser"))
+    implementation(project(":idl-models"))
 
     implementation(project(":logging"))
     implementation(project(":panic"))
