+++ idl/idl-parser/build.gradle
@@ -8,7 +8,7 @@ dependencies {
 
     implementation "org.antlr:antlr4-runtime:${gradle.antlr4_version}"
 
-    implementation(project(":idl:idl-declarations"))
+    implementation(project(":idl-declarations"))
     implementation(project(":ast-common"))
 }
