+++ idl/idl-parser/src/org/jetbrains/dukat/idlParser/visitors/DefinitionVisitor.kt
@@ -1,0 +1,26 @@
+package org.jetbrains.dukat.idlParser.visitors
+
+import org.antlr.webidl.WebIDLBaseVisitor
+import org.antlr.webidl.WebIDLParser
+import org.jetbrains.dukat.idlDeclarations.IDLAttributeDeclaration
+import org.jetbrains.dukat.idlDeclarations.IDLDeclaration
+import org.jetbrains.dukat.idlDeclarations.IDLInterfaceDeclaration
+import org.jetbrains.dukat.idlParser.getName
+
+internal class DefinitionVisitor : WebIDLBaseVisitor<IDLDeclaration>() {
+    private val attributes: MutableList<IDLAttributeDeclaration> = mutableListOf()
+
+    override fun visitAttributeRest(ctx: WebIDLParser.AttributeRestContext): IDLDeclaration {
+        with (AttributeVisitor()) {
+            visit(ctx)
+            this@DefinitionVisitor.attributes.add(visitAttributeRest(ctx))
+        }
+        return IDLInterfaceDeclaration("", listOf())
+    }
+
+    override fun visitInterface_(ctx: WebIDLParser.Interface_Context): IDLDeclaration {
+        val name = getName(ctx)
+        visitChildren(ctx)
+        return IDLInterfaceDeclaration(name, attributes)
+    }
+}
