+++ idl/idl-parser/src/org/jetbrains/dukat/idlParser/visitors/ModuleVisitor.kt
@@ -1,0 +1,13 @@
+package org.jetbrains.dukat.idlParser.visitors
+
+import org.antlr.webidl.WebIDLBaseVisitor
+import org.antlr.webidl.WebIDLParser
+import org.jetbrains.dukat.idlDeclarations.IDLDeclaration
+
+internal class ModuleVisitor(val declarations: MutableList<IDLDeclaration>) : WebIDLBaseVisitor<Unit>() {
+
+    override fun visitDefinition(ctx: WebIDLParser.DefinitionContext?) {
+        declarations.add(DefinitionVisitor().visitDefinition(ctx))
+    }
+
+}
