+++ settings.gradle
@@ -41,9 +41,9 @@ include 'ast-common'
 include 'ast-model'
 include 'cli'
 include 'compiler'
-include 'idl:idl-declarations'
-include 'idl:idl-models'
-include 'idl:idl-parser'
+include 'idl-declarations'
+include 'idl-models'
+include 'idl-parser'
 include 'itertools'
 include 'interop'
 include 'interop-graal'
@@ -59,3 +59,7 @@ include 'ts'
 include 'tsmodel'
 include 'tsinterop'
 include 'translator'
+
+project(':idl-declarations').projectDir = "$rootDir/idl/idl-declarations" as File
+project(':idl-models').projectDir = "$rootDir/idl/idl-models" as File
+project(':idl-parser').projectDir = "$rootDir/idl/idl-parser" as File
